Feature separation in Infrastructure project
