+++ templates/MicroserviceTemplate.Aspire/src/MyService.Infrastructure/Common/Messaging/RabbitMqPublisher.cs
@@ -4,7 +4,7 @@ using Microsoft.Extensions.Configuration;
 using MyService.Domain.Common.Interfaces;
 using RabbitMQ.Client;
 
-namespace MyService.Infrastructure.Messaging;
+namespace MyService.Infrastructure.Common.Messaging;
 
 public class RabbitMqPublisher : IMessagePublisher
 {
