+++ templates/MicroserviceTemplate.Aspire/src/MyService.Infrastructure/Services/OutboxService.cs
@@ -3,7 +3,7 @@ using Microsoft.EntityFrameworkCore;
 using MyService.Domain.Common.Entities;
 using MyService.Infrastructure.Data;
 
-namespace MyService.Infrastructure.Services;
+namespace MyService.Infrastructure.Common.Services;
 
 /// <summary>
 /// Represents an outbox service.
